requirements.txt
runpod==1.7.13
requests==2.32.3
pillow==10.4.0